+++ Editor/Data/ImportRule.cs
@@ -1,4 +1,6 @@
 using System;
+using System.Collections.Generic;
+using Kodlon.AssetRouter.Actions;
 using UnityEditor.Presets;
 
 namespace Kodlon.AssetRouter.Data
@@ -7,5 +9,7 @@ namespace Kodlon.AssetRouter.Data
     public class ImportRule : BaseImportRule
     {
         public Preset preset;
+
+        public List<AssetImportActionAsset> postImportActions = new();
     }
 }